fix(input): shadow rename, sign-conversion casts, default member initializers
Replace ControllerState constructor with default member initializers for
connected and handle. Apply static_cast<std::size_t> to all controller array
indexing operations to resolve sign-conversion warnings. Rename rightStickY to
scrollStickValue to eliminate variable shadowing.

+++ src/input/input_controller.cpp
@@ -10,8 +10,8 @@ void Input::initializeController() {
             SDL_GameController* handle = SDL_GameControllerOpen(i);
             if (!handle) continue;
             SDL_JoystickID id = SDL_JoystickInstanceID(SDL_GameControllerGetJoystick(handle));
-            controllers[slot].handle = handle;
-            controllers[slot].connected = true;
+            controllers[static_cast<std::size_t>(slot)].handle = handle;
+            controllers[static_cast<std::size_t>(slot)].connected = true;
             instanceToSlot[id] = slot;
             ++slot;
         }
